Moving synchronization and threading to iree/base/internal/.
