+++ iree/base/BUILD
@@ -148,90 +148,6 @@ cc_test(
     srcs = ["string_view_test.cc"],
     deps = [
         ":base",
-        ":api",
-        "//iree/testing:gtest",
-        "//iree/testing:gtest_main",
-    ],
-)
-
-cc_library(
-    name = "synchronization",
-    srcs = ["synchronization.c"],
-    hdrs = ["synchronization.h"],
-    deps = [
-        ":api",
-        ":core_headers",
-        ":tracing",
-        "//iree/base/internal",
-    ],
-)
-
-cc_binary(
-    name = "synchronization_benchmark",
-    testonly = True,
-    srcs = ["synchronization_benchmark.cc"],
-    deps = [
-        ":synchronization",
-        "//iree/testing:benchmark_main",
-        "@com_google_benchmark//:benchmark",
-    ],
-)
-
-run_binary_test(
-    name = "synchronization_benchmark_test",
-    args = ["--benchmark_min_time=0"],
-    test_binary = ":synchronization_benchmark",
-)
-
-cc_test(
-    name = "synchronization_test",
-    srcs = ["synchronization_test.cc"],
-    deps = [
-        ":synchronization",
-        "//iree/testing:gtest",
-        "//iree/testing:gtest_main",
-    ],
-)
-
-cc_library(
-    name = "threading",
-    srcs = [
-        "threading.c",
-        "threading_darwin.c",
-        "threading_impl.h",
-        "threading_pthreads.c",
-        "threading_win32.c",
-    ],
-    hdrs = ["threading.h"],
-    deps = [
-        ":api",
-        ":core_headers",
-        ":synchronization",
-        ":tracing",
-        "//build_tools:default_linkopts",
-        "//iree/base/internal",
-    ],
-)
-
-cc_test(
-    name = "threading_benchmark",
-    srcs = ["threading_benchmark.cc"],
-    deps = [
-        ":threading",
-        "//iree/testing:benchmark_main",
-        "@com_google_benchmark//:benchmark",
-    ],
-)
-
-cc_test(
-    name = "threading_test",
-    srcs = [
-        "threading_impl.h",
-        "threading_test.cc",
-    ],
-    deps = [
-        ":synchronization",
-        ":threading",
         "//iree/testing:gtest",
         "//iree/testing:gtest_main",
     ],
