+++ iree/base/CMakeLists.txt
@@ -141,98 +141,6 @@ iree_cc_test(
     "string_view_test.cc"
   DEPS
     ::base
-    ::api
-    iree::testing::gtest
-    iree::testing::gtest_main
-)
-
-iree_cc_library(
-  NAME
-    synchronization
-  HDRS
-    "synchronization.h"
-  SRCS
-    "synchronization.c"
-  DEPS
-    ::api
-    ::core_headers
-    ::tracing
-    iree::base::internal
-  PUBLIC
-)
-
-iree_cc_binary(
-  NAME
-    synchronization_benchmark
-  SRCS
-    "synchronization_benchmark.cc"
-  DEPS
-    ::synchronization
-    benchmark
-    iree::testing::benchmark_main
-  TESTONLY
-)
-
-iree_run_binary_test(
-  NAME
-    synchronization_benchmark_test
-  TEST_BINARY
-    ::synchronization_benchmark
-  ARGS
-    "--benchmark_min_time=0"
-)
-
-iree_cc_test(
-  NAME
-    synchronization_test
-  SRCS
-    "synchronization_test.cc"
-  DEPS
-    ::synchronization
-    iree::testing::gtest
-    iree::testing::gtest_main
-)
-
-iree_cc_library(
-  NAME
-    threading
-  HDRS
-    "threading.h"
-  SRCS
-    "threading.c"
-    "threading_darwin.c"
-    "threading_impl.h"
-    "threading_pthreads.c"
-    "threading_win32.c"
-  DEPS
-    ::api
-    ::core_headers
-    ::synchronization
-    ::tracing
-    iree::base::internal
-  PUBLIC
-)
-
-iree_cc_test(
-  NAME
-    threading_benchmark
-  SRCS
-    "threading_benchmark.cc"
-  DEPS
-    ::threading
-    benchmark
-    iree::testing::benchmark_main
-)
-
-iree_cc_test(
-  NAME
-    threading_test
-  SRCS
-    "threading_impl.h"
-    "threading_test.cc"
-  DEPS
-    ::synchronization
-    ::threading
     iree::testing::gtest
     iree::testing::gtest_main
 )
