+++ iree/base/internal/CMakeLists.txt
@@ -58,8 +58,8 @@ iree_cc_library(
     "atomic_slist.c"
   DEPS
     ::internal
+    ::synchronization
     iree::base::core_headers
-    iree::base::synchronization
   PUBLIC
 )
 
@@ -85,8 +85,8 @@ iree_cc_library(
   DEPS
     ::file_path
     ::internal
+    ::threading
     iree::base::core_headers
-    iree::base::threading
     iree::base::tracing
   PUBLIC
 )
@@ -230,6 +230,97 @@ iree_cc_library(
     iree::base::core_headers
     iree::base::logging
   PUBLIC
+)
+
+iree_cc_library(
+  NAME
+    synchronization
+  HDRS
+    "synchronization.h"
+  SRCS
+    "synchronization.c"
+  DEPS
+    ::internal
+    iree::base
+    iree::base::core_headers
+    iree::base::tracing
+  PUBLIC
+)
+
+iree_cc_binary(
+  NAME
+    synchronization_benchmark
+  SRCS
+    "synchronization_benchmark.cc"
+  DEPS
+    ::synchronization
+    benchmark
+    iree::testing::benchmark_main
+  TESTONLY
+)
+
+iree_run_binary_test(
+  NAME
+    "synchronization_benchmark_test"
+  ARGS
+    "--benchmark_min_time=0"
+  TEST_BINARY
+    ::synchronization_benchmark
+)
+
+iree_cc_test(
+  NAME
+    synchronization_test
+  SRCS
+    "synchronization_test.cc"
+  DEPS
+    ::synchronization
+    iree::testing::gtest
+    iree::testing::gtest_main
+)
+
+iree_cc_library(
+  NAME
+    threading
+  HDRS
+    "threading.h"
+  SRCS
+    "threading.c"
+    "threading_darwin.c"
+    "threading_impl.h"
+    "threading_pthreads.c"
+    "threading_win32.c"
+  DEPS
+    ::internal
+    ::synchronization
+    iree::base
+    iree::base::core_headers
+    iree::base::tracing
+  PUBLIC
+)
+
+iree_cc_test(
+  NAME
+    threading_benchmark
+  SRCS
+    "threading_benchmark.cc"
+  DEPS
+    ::threading
+    benchmark
+    iree::testing::benchmark_main
+)
+
+iree_cc_test(
+  NAME
+    threading_test
+  SRCS
+    "threading_impl.h"
+    "threading_test.cc"
+  DEPS
+    ::synchronization
+    ::threading
+    iree::testing::gtest
+    iree::testing::gtest_main
 )
 
 iree_cc_library(
