+++ iree/base/internal/dynamic_library_win32.c
@@ -17,8 +17,8 @@
 #include "iree/base/internal/atomics.h"
 #include "iree/base/internal/dynamic_library.h"
 #include "iree/base/internal/file_path.h"
+#include "iree/base/internal/threading.h"
 #include "iree/base/target_platform.h"
-#include "iree/base/threading.h"
 #include "iree/base/tracing.h"
 
 #if defined(IREE_PLATFORM_WINDOWS)
