+++ iree/base/internal/threading_win32.c
@@ -13,15 +13,15 @@
 // limitations under the License.
 
 // NOTE: must be first to ensure that we can define settings for all includes.
-#include "iree/base/threading_impl.h"
+#include "iree/base/internal/threading_impl.h"
 
 #if defined(IREE_PLATFORM_WINDOWS)
 
 #include <stdio.h>
 
 #include "iree/base/internal/atomics.h"
-#include "iree/base/threading.h"
-#include "iree/base/threading_impl.h"
+#include "iree/base/internal/threading.h"
+#include "iree/base/internal/threading_impl.h"
 #include "iree/base/tracing.h"
 
 // Great documentation:
