+++ iree/hal/CMakeLists.txt
@@ -58,8 +58,8 @@ iree_cc_library(
     iree::base
     iree::base::core_headers
     iree::base::internal
-    iree::base::synchronization
-    iree::base::threading
+    iree::base::internal::synchronization
+    iree::base::internal::threading
     iree::base::tracing
   PUBLIC
 )
