+++ iree/hal/local/sync_semaphore.h
@@ -16,7 +16,7 @@
 #define IREE_HAL_LOCAL_SYNC_SEMAPHORE_H_
 
 #include "iree/base/api.h"
-#include "iree/base/synchronization.h"
+#include "iree/base/internal/synchronization.h"
 #include "iree/hal/api.h"
 
 #ifdef __cplusplus
