+++ iree/hal/vulkan/BUILD
@@ -96,7 +96,7 @@ cc_library(
         "//iree/base:flatcc",
         "//iree/base:logging",
         "//iree/base:status",
-        "//iree/base:synchronization",
+        "//iree/base/internal:synchronization",
         "//iree/base:tracing",
         "//iree/base/internal",
         "//iree/hal",
