+++ iree/hal/vulkan/command_queue.h
@@ -17,8 +17,8 @@
 
 #include <string>
 
+#include "iree/base/internal/synchronization.h"
 #include "iree/base/status.h"
-#include "iree/base/synchronization.h"
 #include "iree/hal/api.h"
 #include "iree/hal/vulkan/dynamic_symbols.h"
 #include "iree/hal/vulkan/handle_util.h"
