+++ iree/hal/vulkan/handle_util.h
@@ -28,8 +28,8 @@
 #include "iree/hal/vulkan/vulkan_headers.h"
 // clang-format on
 
+#include "iree/base/internal/synchronization.h"
 #include "iree/base/status.h"
-#include "iree/base/synchronization.h"
 #include "iree/hal/vulkan/dynamic_symbols.h"
 #include "iree/hal/vulkan/extensibility_util.h"
 #include "iree/hal/vulkan/status_util.h"
