+++ iree/hal/vulkan/serializing_command_queue.h
@@ -23,8 +23,8 @@
 #include <string>
 #include <vector>
 
+#include "iree/base/internal/synchronization.h"
 #include "iree/base/status.h"
-#include "iree/base/synchronization.h"
 #include "iree/hal/api.h"
 #include "iree/hal/vulkan/command_queue.h"
 #include "iree/hal/vulkan/dynamic_symbols.h"
