+++ iree/task/BUILD
@@ -62,12 +62,12 @@ cc_library(
     deps = [
         "//iree/base",
         "//iree/base:core_headers",
-        "//iree/base:synchronization",
-        "//iree/base:threading",
         "//iree/base:tracing",
         "//iree/base/internal",
         "//iree/base/internal:atomic_slist",
         "//iree/base/internal:prng",
+        "//iree/base/internal:synchronization",
+        "//iree/base/internal:threading",
         "//iree/base/internal:wait_handle",
         "@cpuinfo",
     ],
