+++ iree/task/queue.h
@@ -16,7 +16,7 @@
 #define IREE_TASK_QUEUE_H_
 
 #include "iree/base/api.h"
-#include "iree/base/synchronization.h"
+#include "iree/base/internal/synchronization.h"
 #include "iree/task/list.h"
 #include "iree/task/task.h"
