+++ iree/task/scope.h
@@ -17,7 +17,7 @@
 
 #include "iree/base/api.h"
 #include "iree/base/internal/atomics.h"
-#include "iree/base/synchronization.h"
+#include "iree/base/internal/synchronization.h"
 #include "iree/base/tracing.h"
 #include "iree/task/task.h"
