+++ iree/task/worker.h
@@ -16,8 +16,8 @@
 #define IREE_TASK_WORKER_H_
 
 #include "iree/base/internal/prng.h"
-#include "iree/base/synchronization.h"
-#include "iree/base/threading.h"
+#include "iree/base/internal/synchronization.h"
+#include "iree/base/internal/threading.h"
 #include "iree/base/tracing.h"
 #include "iree/task/affinity_set.h"
 #include "iree/task/executor.h"
